requirements.txt
praw==7.6.0
pymongo==4.5.0
python-dotenv==1.0.0